+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/organisationunit/OrganisationUnit.java
@@ -429,7 +429,7 @@ public class OrganisationUnit
         return null;
     }
 
-    public Integer getGroupIdInGroupSet( OrganisationUnitGroupSet groupSet )
+    public Long getGroupIdInGroupSet( OrganisationUnitGroupSet groupSet )
     {
         final OrganisationUnitGroup group = getGroupInGroupSet( groupSet );
