+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/validation/ValidationResultStore.java
@@ -44,7 +44,7 @@ public interface ValidationResultStore
 {
     List<ValidationResult> getAllUnreportedValidationResults();
 
-    ValidationResult getById( int id );
+    ValidationResult getById( long id );
 
     List<ValidationResult> query( ValidationResultQuery query );
