+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/metadata/version/hibernate/HibernateMetadataVersionStore.java
@@ -46,7 +46,7 @@ public class HibernateMetadataVersionStore
     implements MetadataVersionStore
 {
     @Override
-    public MetadataVersion getVersionByKey( int key )
+    public MetadataVersion getVersionByKey( long key )
     {
         CriteriaBuilder builder = getCriteriaBuilder();
