+++ dhis-2/dhis-services/dhis-service-dxf2/src/main/java/org/hisp/dhis/dxf2/pdfform/DefaultPdfDataEntryFormService.java
@@ -35,7 +35,6 @@ import java.util.Calendar;
 import java.util.Collection;
 import java.util.List;
 
-import org.hisp.dhis.api.util.DateUtils;
 import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.common.ValueType;
 import org.hisp.dhis.dataelement.DataElement;
@@ -60,6 +59,7 @@ import org.hisp.dhis.period.YearlyPeriodType;
 import org.hisp.dhis.program.ProgramStage;
 import org.hisp.dhis.program.ProgramStageSection;
 import org.hisp.dhis.program.ProgramStageService;
+import org.hisp.dhis.util.DateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.lowagie.text.Chunk;
@@ -437,7 +437,7 @@ public class DefaultPdfDataEntryFormService
                 OptionSet optionSet = dataElement.getOptionSet();
 
                 String strFieldLabel = PdfDataEntryFormUtil.LABELCODE_DATAENTRYTEXTFIELD
-                    + Integer.toString( dataElement.getId() )
+                    + Long.toString( dataElement.getId() )
                     // + "_" + Integer.toString(programStageId) + "_" +
                     // Integer.toString(rowNo);
                     + "_" + Integer.toString( rowNo );
