+++ dhis-2/dhis-services/dhis-service-mobile/src/main/java/org/hisp/dhis/api/mobile/FacilityReportingService.java
@@ -46,9 +46,9 @@ public interface FacilityReportingService
 
     List<DataSet> getMobileDataSetsForUnit( OrganisationUnit unit, String localeString );
 
-    DataSet getDataSet( int id );
+    DataSet getDataSet( long id );
 
-    DataSet getDataSetForLocale( int dataSetId, Locale locale );
+    DataSet getDataSetForLocale( long dataSetId, Locale locale );
 
     void saveDataSetValues( OrganisationUnit unit, DataSetValue dataSetValue )
         throws NotAllowedException;
